requirements.txt
alembic==1.6.2
amqp==5.0.6
attrs==21.1.0
base58==2.1.0
billiard==3.6.4.0
bitarray==1.2.2
bscscan-python==1.0.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cytoolz==0.11.0
docker==5.0.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
etherscan==0.3.0
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Web3==0.1.1
greenlet==1.1.0
heroku==0.1.4
hexbytes==0.2.1
idna==2.10
importlib-metadata==4.0.1
ipfshttpclient==0.7.0a1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kombu==5.0.2
lru-dict==1.1.7
Mako==1.1.4
MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.8.0
numpy==1.20.2
pandas==1.2.4
parsimonious==0.8.1
prompt-toolkit==3.0.18
protobuf==3.15.8
pycryptodome==3.10.1
pyrsistent==0.17.3
python-dateutil==1.5
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-cache==0.6.3
rlp==2.0.1
rq==1.8.1
six==1.16.0
SQLAlchemy==1.4.14
timestamp==0.0.1
toolz==0.11.1
typing-extensions==3.10.0.0
url-normalize==1.4.3
urllib3==1.26.4
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.19.0
websocket-client==1.0.0
websockets==8.1
Werkzeug==1.0.1
zipp==3.4.1
